requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
Django==3.2.7
django-cors-headers==3.6.0
django-crontab==0.7.1
django-jsonfield==1.4.1
PyJWT==1.7.1
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
uWSGI==2.0.18
